tweeks engaged

+++ BobbysBurden/src/actions/PushGuestBedButtonAction.cpp
@@ -15,7 +15,9 @@ void PushGuestBedButtonAction::perform()
 	tilePosition.x -= 0;
 	tilePosition.y += 2;
 
-	const auto& keyObject = m_parent->parentScene()->addGameObject("BED_STAIRS_BLOCK", nullptr, GameLayer::BACKGROUND_3, tilePosition.x, tilePosition.y);
+	const auto& keyObject = m_parent->parentScene()->
+		addGameObject("BED_STAIRS_BLOCK", nullptr, GameLayer::BACKGROUND_3, 
+			static_cast<float>(tilePosition.x), static_cast<float>(tilePosition.y));
 
 	//disable this puzzle object so that the action can only be triggered once
 	m_parent->setRemoveFromWorld(true);
